requirements.txt
pytorch-pretrained-biggan
transformers
matplotlib
nltk
Pillow
tqdm
jupyter
scipy
tensorflow
keytotext
cStringIO